Move to sim folder, and reorg (no API breaks)
